+++ packages/accounts-controller/src/AccountsController.ts
@@ -162,7 +162,7 @@ export type AccountsControllerAccountAddedEvent = {
 
 export type AccountsControllerAccountRemovedEvent = {
   type: `${typeof controllerName}:accountRemoved`;
-  payload: [AccountId];
+  payload: [{ id: string; address: string }];
 };
 
 export type AccountsControllerAccountRenamedEvent = {
@@ -862,8 +862,12 @@ export class AccountsController extends BaseController<
     }
 
     // Diff that we will use to publish events afterward.
+    type diffItem = {
+      id: string;
+      address: string;
+    };
     const diff = {
-      removed: [] as string[],
+      removed: [] as diffItem[],
       added: [] as InternalAccount[],
     };
 
@@ -874,7 +878,10 @@ export class AccountsController extends BaseController<
         for (const account of patch.removed) {
           delete internalAccounts.accounts[account.id];
 
-          diff.removed.push(account.id);
+          diff.removed.push({
+            id: account.id,
+            address: account.address,
+          });
         }
 
         for (const added of patch.added) {
@@ -921,8 +928,11 @@ export class AccountsController extends BaseController<
     });
 
     // Now publish events
-    for (const id of diff.removed) {
-      this.messagingSystem.publish('AccountsController:accountRemoved', id);
+    for (const item of diff.removed) {
+      this.messagingSystem.publish('AccountsController:accountRemoved', {
+        id: item.id,
+        address: item.address,
+      });
     }
 
     for (const account of diff.added) {