+++ packages/assets-controllers/src/MultichainAssetsController/MultichainAssetsController.test.ts
@@ -534,7 +534,10 @@ describe('MultichainAssetsController', () => {
       assetsMetadata: mockGetMetadataReturnValue.assets,
     });
     // Remove an EVM account
-    messenger.publish('AccountsController:accountRemoved', mockEthAccount.id);
+    messenger.publish('AccountsController:accountRemoved', {
+      id: mockEthAccount.id,
+      address: mockEthAccount.address,
+    });
 
     await advanceTime({ clock, duration: 1 });
 
@@ -580,10 +583,10 @@ describe('MultichainAssetsController', () => {
       assetsMetadata: mockGetMetadataReturnValue.assets,
     });
     // Remove the added solana account
-    messenger.publish(
-      'AccountsController:accountRemoved',
-      mockSolanaAccount.id,
-    );
+    messenger.publish('AccountsController:accountRemoved', {
+      id: mockSolanaAccount.id,
+      address: mockSolanaAccount.address,
+    });
 
     await advanceTime({ clock, duration: 1 });